+++ spec/spec_helper.rb
@@ -1,3 +1,9 @@
+require 'yaml'
+begin
+  require 'psych'
+rescue LoadError
+end
+
 RSpec::Matchers.define :include_method do |expected|
   match do |actual|
     actual.map { |m| m.to_s }.include?(expected.to_s)